Refactor reproduction tests to align with project style
Split tests into runners extending SimpleBaseTest and sample test cases, ensuring proper isolation and style consistency.
