+++ testng-core/src/test/java/test/dataprovider/issue3236/IssueTest.java
@@ -1,45 +1,23 @@
 package test.dataprovider.issue3236;
 
-import java.util.concurrent.atomic.AtomicInteger;
-import org.testng.Assert;
-import org.testng.IRetryAnalyzer;
-import org.testng.ITestResult;
-import org.testng.annotations.DataProvider;
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.testng.TestListenerAdapter;
+import org.testng.TestNG;
 import org.testng.annotations.Test;
+import test.SimpleBaseTest;
 
-public class IssueTest {
+public class IssueTest extends SimpleBaseTest {
 
-  private static final AtomicInteger invocationCount = new AtomicInteger(0);
-  private static long lastValue = 0;
+  @Test
+  public void verifyDataProviderIsReExecuted() {
+    TestNG tng = create(SampleTestCase.class);
+    TestListenerAdapter tla = new TestListenerAdapter();
+    tng.addListener(tla);
+    tng.run();
 
-  @Test(dataProvider = "dp", retryAnalyzer = MyRetry.class)
-  public void testMethod(long value) {
-    int count = invocationCount.getAndIncrement();
-    if (count == 0) {
-        lastValue = value;
-        throw new RuntimeException("Fail first time to trigger retry");
-    } else {
-        // Retry
-        Assert.assertNotEquals(value, lastValue, "DataProvider should have been re-executed and returned a new value");
-    }
-  }
-
-  @DataProvider(name = "dp", cacheDataForTestRetries = false)
-  public Object[][] getData() {
-    return new Object[][] {{ System.nanoTime() }};
-  }
-
-  public static class MyRetry implements IRetryAnalyzer {
-    private int retryCount = 0;
-    private static final int maxRetryCount = 1;
-
-    @Override
-    public boolean retry(ITestResult result) {
-      if (retryCount < maxRetryCount) {
-        retryCount++;
-        return true;
-      }
-      return false;
-    }
+    assertThat(tla.getFailedTests()).isEmpty();
+    assertThat(tla.getPassedTests()).hasSize(1);
+    assertThat(tla.getSkippedTests()).hasSize(1);
   }
 }
