+++ testng-core/src/test/java/test/dataprovider/issue3237/CacheTest.java
@@ -1,43 +1,23 @@
 package test.dataprovider.issue3237;
 
-import java.util.UUID;
-import java.util.concurrent.atomic.AtomicInteger;
-import org.testng.Assert;
-import org.testng.IRetryAnalyzer;
-import org.testng.ITestResult;
-import org.testng.annotations.DataProvider;
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.testng.TestListenerAdapter;
+import org.testng.TestNG;
 import org.testng.annotations.Test;
+import test.SimpleBaseTest;
 
-public class CacheTest {
+public class CacheTest extends SimpleBaseTest {
 
-  private final AtomicInteger invocationCount = new AtomicInteger(0);
-  private String currentUuid;
+  @Test
+  public void verifyFreshDataOnRetry() {
+    TestNG tng = create(SampleCacheTestCase.class);
+    TestListenerAdapter tla = new TestListenerAdapter();
+    tng.addListener(tla);
+    tng.run();
 
-  @Test(dataProvider = "dp", retryAnalyzer = MyRetry.class)
-  public void testMethod(String uuid) {
-    // Verify that fresh UUID is generated on each retry (no caching)
-    // If bug exists: DP runs (updates currentUuid), but Test receives OLD uuid. Expect failure.
-    // If fixed: DP runs (updates currentUuid), Test receives NEW uuid. Expect pass.
-    Assert.assertEquals(uuid, currentUuid, "Received stale data from DataProvider!");
-
-    // Fail on first invocation to trigger retry, pass on second
-    if (invocationCount.getAndIncrement() < 1) {
-      throw new RuntimeException("Failed for " + uuid);
-    }
-  }
-
-  @DataProvider(name = "dp", cacheDataForTestRetries = false)
-  public Object[][] getData() {
-    currentUuid = UUID.randomUUID().toString();
-    return new Object[][] {{ currentUuid }};
-  }
-
-  public static class MyRetry implements IRetryAnalyzer {
-    private int count = 0;
-
-    @Override
-    public boolean retry(ITestResult result) {
-      return count++ < 1;
-    }
+    assertThat(tla.getFailedTests()).isEmpty();
+    assertThat(tla.getPassedTests()).hasSize(1);
+    assertThat(tla.getSkippedTests()).hasSize(1);
   }
 }
